requirements/base.txt
asgiref==3.11.1
Django==6.0.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
sqlparse==0.5.5
asgiref==3.11.1
Django==6.0.2
django-ratelimit==4.1.0
python-decouple==3.8
attrs==25.4.0
django-debug-toolbar==6.2.0
django-extensions==4.1
django-redis==6.0.0
drf-spectacular==0.29.0
Faker==40.4.0
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
logging==0.4.9.6
pillow==12.1.1
PyJWT==2.11.0
PyYAML==6.0.3
redis==7.1.1
referencing==0.37.0
rpds-py==0.30.0
typing_extensions==4.15.0
uritemplate==4.2.0
